Remove really jbrowse-web specific stuff
