--- products/jbrowse-react-app/src/factoryReset.ts
@@ -1,4 +1,0 @@
-export default async function factoryReset() {
-  // @ts-expect-error
-  window.location = window.location.pathname
-}
