--- products/jbrowse-react-app/src/sessionModelFactory.ts
@@ -1,5 +1,0 @@
-import oldSessionModelFactory from './sessionModel'
-
-/** @deprecated moved to ./sessionModel */
-const sessionModelFactory = oldSessionModelFactory
-export default sessionModelFactory
