+++ products/jbrowse-react-app/src/version.js
@@ -1,1 +1,1 @@
-export const version = '2.3.3'
+export const version = '2.5.0'
