+++ products/jbrowse-react-app/stories/JBrowseReactApp.stories.tsx
@@ -1,0 +1,3 @@
+export { BasicExample } from './examples'
+
+export default { title: 'Source code for examples' }
