+++ products/jbrowse-react-app/stories/examples/index.ts
@@ -1,0 +1,1 @@
+export * from './BasicExample'
