+++ products/jbrowse-react-app/stories/util.ts
@@ -1,0 +1,14 @@
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export function addRelativeUris(config: any, baseUri: string) {
+  if (typeof config === 'object') {
+    for (const key of Object.keys(config)) {
+      if (typeof config[key] === 'object') {
+        addRelativeUris(config[key], baseUri)
+      } else if (key === 'uri') {
+        if (!config.baseUri) {
+          config.baseUri = baseUri
+        }
+      }
+    }
+  }
+}
